requirements.txt
qrcode[pil]==7.4.2
Pillow==10.0.1
Flask==2.3.3
Werkzeug==2.3.7
Flask==2.3.3
Werkzeug==2.3.7
qrcode[pil]==7.4.2
Pillow==10.0.1
gunicorn==21.2.0